Moved parse to module, replaced plot dropdowns with listboxes
