+++ parse/__init__.py
@@ -22,189 +22,14 @@
 
 import types
 import time
+from parse.clean import *
+from parse.get import *
+from parse.prep import *
 
 
-def get_pid(config_info, match):
-    # Gets the player's ID for the match
-    pid = "Unknown"
-    n_players = len(match["participantIdentities"])
-    for ii in range(n_players):
-        if (str(match["participantIdentities"][ii]["player"]["accountId"]) == str(config_info["AccountID"])):
-            pid = ii
-    return pid
-
-
-def get_role_pretty(config_info, match):
-
-    pid = get_pid(config_info, match)
-
-    try:
-        role = str(match["participants"][pid]["timeline"]["role"])
-        lane = str(match["participants"][pid]["timeline"]["lane"])
-    except:
-        role = "Unknown"
-        lane = "Unknown"
-
-    if lane.lower() == "middle" or lane.lower() == "mid":
-        role_pretty = "Middle"
-    elif lane.lower() == "top":
-        role_pretty = "Top"
-    elif lane.lower() == "jungle" or role.lower() == "jungle":
-        role_pretty = "Jungle"
-    elif lane.lower() == "bottom" or lane.lower() == "bot":
-        if "carry" in role.lower():
-            role_pretty = "Bot/ADC"
-        elif "sup" in role.lower():
-            role_pretty = "Support"
-        else:
-            role_pretty = "Bot (Other)"
-    else:
-        role_pretty = "Unknown"
-
-    return role_pretty
-
-
-def get_kda(config_info, match):
-    pid = get_pid(config_info, match)
-
-    kills = match["participants"][pid]["stats"]["kills"]
-    deaths = match["participants"][pid]["stats"]["deaths"]
-    assists = match["participants"][pid]["stats"]["assists"]
-
-    try:
-        kda = str(round(
-            (int(kills) + int(assists))
-            / int(deaths)
-        , 1))
-    except ZeroDivisionError:
-        kda = "Perfect"
-    except TypeError:
-        kda = "Unknown"
-
-    return kda
-
-
-def get_oid(config_info, match):
-    # Gets the player's ID for the match
-    pid = get_pid(config_info, match)
-    oid = "Unknown"
-    n_players = len(match["participantIdentities"])
-
-    lane = str(match["participants"][pid]["timeline"]["lane"])
-    role = str(match["participants"][pid]["timeline"]["role"])
-
-    for ii in range(n_players):
-        if (str(match["participants"][ii]["timeline"]["lane"]) == str(lane) and
-                    str(match["participants"][ii]["timeline"]["role"]) == str(role) and ii != pid):
-            oid = ii
-
-    return oid
-
-
-def get_count(config_info, match):
-    return 1
-
-
-def get_teammates(config_info, match):
-    teammates = []
-    pid = get_pid(config_info, match)
-
-    n_players = len(match["participantIdentities"])
-    for ii in range(n_players):
-        if str(match["participants"][ii]["teamId"]) == str(match["participants"][pid]["teamId"]) and ii != pid:
-            teammates.append(
-                match["participantIdentities"][ii]["player"]["summonerName"]
-            )
-
-    if len(teammates) == 0:
-        teammates = ["Unknown"]
-    teammates = sorted(teammates)
-
-    return teammates
-
-
-def get_opponents(config_info, match):
-    opponents = []
-    pid = get_pid(config_info, match)
-
-    n_players = len(match["participantIdentities"])
-    for ii in range(n_players):
-        if str(match["participants"][ii]["teamId"]) != str(match["participants"][pid]["teamId"]):
-            opponents.append(
-                match["participantIdentities"][ii]["player"]["summonerName"]
-            )
-
-    if len(opponents) == 0:
-        opponents = ["Unknown"]
-    opponents = sorted(opponents)
-
-    return opponents
-
-
-def get_non_player_champs(config_info, match):
-    non_player_champs = []
-    pid = get_pid(config_info, match)
-    n_players = len(match["participantIdentities"])
-
-    for ii in range(n_players):
-        if ii != pid:
-            champ_id = match["participants"][ii]["championId"]
-            champ = clean_champion(config_info, champ_id)
-            non_player_champs.append(champ)
-
-    if len(non_player_champs) == 0:
-        non_player_champs = ["Unknown"]
-
-    return non_player_champs
-
-
-def get_num_teammates(config_info, match_data):
-    num_teammates_list = []
-
-    return num_teammates_list
 
 # TODO: remove testing code
 testing=0
-
-
-def get_teammate_champs(config_info, match):
-    # TODO: check this and enemy version
-    teammate_champs = []
-    pid = get_pid(config_info, match)
-    n_players = len(match["participantIdentities"])
-
-    for ii in range(n_players):
-        if str(match["participants"][ii]["teamId"]) == str(match["participants"][pid]["teamId"]) and ii != pid:
-            champ_id = match["participants"][ii]["championId"]
-            champ = clean_champion(config_info, champ_id)
-            teammate_champs.append(champ)
-
-    if len(teammate_champs) == 0:
-        teammate_champs = ["Unknown"]
-    return teammate_champs
-
-
-def get_enemy_champs(config_info, match):
-    enemy_champs = []
-    pid = get_pid(config_info, match)
-    n_players = len(match["participantIdentities"])
-
-    for ii in range(n_players):
-        if str(match["participants"][ii]["teamId"]) != str(match["participants"][pid]["teamId"]):
-            champ_id = match["participants"][ii]["championId"]
-            champ = clean_champion(config_info, champ_id)
-            enemy_champs.append(champ)
-
-    if len(enemy_champs) == 0:
-        enemy_champs = ["Unknown"]
-    return enemy_champs
-
-
-
-
-
-
-
 if testing:
     import json
     with open("Configuration.json", "r") as file:
@@ -215,184 +40,9 @@ if testing:
     # print(json.dumps(match["participants"][pid], indent=4))
 # testing = 1
 
-
-
-
-
-
-
-
-
-
-def yvar_vs_count_in_lists(list_of_lists, y_var_list):
-    # total_count = sum(sublist.count("TEAMMATENAME") for sublist in listoflists)
-    return
-
-
-
-def get_items(config_info, match):
-    items = []
-    pid = get_pid(config_info, match)
-
-    for ii in range(7):
-        new_item = match["participants"][pid]["stats"]["item" + str(ii)]
-        if new_item != 0:
-            items += [clean_item(config_info, new_item)]
-
-    items = list(set(items))  # dump duplicate purchases
-
-    return items
-
-
-def get_gold_diffs(config_info, match):
-    gold_diffs_dict = {}
-    pid = get_pid(config_info, match)
-    oid = get_oid(config_info, match)
-
-    try:
-        player_gold_deltas = match["participants"][pid]["timeline"]["goldPerMinDeltas"]
-        opponent_gold_deltas = match["participants"][oid]["timeline"]["goldPerMinDeltas"]
-        gold_diff_keys = list(player_gold_deltas.keys())
-        for kk in gold_diff_keys:
-            gold_diffs_dict[kk] = float(player_gold_deltas[kk]) - float(opponent_gold_deltas[kk])
-
-    except (KeyError, ValueError):
-        pass
-
-    return gold_diffs_dict
-
-
-def get_fractions(config_info, match):
-    fractions_dict = {
-        "goldEarned": "Unknown",
-        "totalDamageDealt": "Unknown",
-        "totalDamageDealtToChampions": "Unknown",
-        "damageDealtToObjectives": "Unknown",
-        "magicDamageDealt": "Unknown",
-        "magicDamageDealtToChampions": "Unknown",
-        "physicalDamageDealt": "Unknown",
-        "physicalDamageDealtToChampions": "Unknown",
-        "trueDamageDealt": "Unknown",
-        "trueDamageDealtToChampions": "Unknown",
-        "totalDamageTaken": "Unknown",
-        "damageSelfMitigated": "Unknown",
-    }
-
-    pid = get_pid(config_info, match)
-    n_players = len(match["participantIdentities"])
-    mitigated_taken_player = 0
-    mitigated_taken_team = 0
-
-    for kk in list(fractions_dict.keys()):
-        try:
-            player_share = 0
-            teammates_share = 0
-            for ii in range(n_players):
-                if ii != pid and str(match["participants"][ii]["teamId"]) == str(match["participants"][pid]["teamId"]):
-                    teammates_share += float(match["participants"][ii]["stats"][kk])
-                elif ii == pid:
-                    player_share = float(match["participants"][ii]["stats"][kk])
-
-            fractions_dict[kk] = player_share / (player_share + teammates_share)
-
-            if kk == "damageSelfMitigated" or kk == "totalDamageTaken":
-                mitigated_taken_player += player_share
-                mitigated_taken_team += (player_share + teammates_share)
-        except (KeyError, ValueError, NameError, ZeroDivisionError) as e:
-            pass
-    try:
-        fractions_dict["damageReceivedTotal"] = mitigated_taken_player / mitigated_taken_team
-        if fractions_dict["damageReceivedTotal"] == fractions_dict["totalDamageTaken"]:
-            raise KeyError  # Don't include games that are missing the damage mitigated stat
-    except (ZeroDivisionError, KeyError):
-        fractions_dict["damageReceivedTotal"] = "Unknown"
-
-    return fractions_dict
-
-
-def clean_game_duration(config_info, game_duration):
-    match_length = round(float(game_duration) / 60, 2)  # match length in minutes
-    return match_length
-
-
-def clean_team(config_info, team):
-    if int(team) == 100:
-        map_side = "Blue"
-    elif int(team) == 200:
-        map_side = "Red"
-    else:
-        map_side = "Other"
-    return map_side
-
-
-def clean_binary(config_info, bin_var):
-    if bin_var == 1 or bin_var == "True":
-        clean_var = 1
-    elif bin_var == 0 or bin_var == "False":
-        clean_var = 0
-    else:
-        clean_var = "Unknown"
-    return clean_var
-
-
-def clean_season(config_info, season_id):
-    try:
-        season_clean = config_info["seasons.gameconstants"][str(season_id)]
-    except:
-        season_clean = "Unknown"
-    return season_clean
-
-
-def clean_queue(config_info, queue_id):
-    queue_clean = "Unknown"
-    try:
-        for queues_key in list(config_info["queues.gameconstants"].keys()):
-            queues_list = queues_key.split("&&")
-            if str(queue_id) in queues_list:
-                queue_clean = config_info["queues.gameconstants"][queues_key]
-    except:
-        queue_clean = "Unknown"
-    return queue_clean
-
-
-def clean_map(config_info, map_id):
-    try:
-        map_clean = config_info["maps.gameconstants"][str(map_id)]
-    except:
-        map_clean = "Unknown"
-    return map_clean
-
-
-def clean_champion(config_info, champ_id):
-    try:
-        champ_name = config_info["champion"][str(champ_id)]
-    except:
-        champ_name = "Unknown"
-    return champ_name
-
-
-def clean_summoner_spell(config_info, spell_id):
-    try:
-        spell_name = config_info["summoner"][str(spell_id)]
-    except:
-        spell_name = "Unknown"
-    return spell_name
-
-
-def clean_item(config_info, item_id):
-    try:
-        item_name = config_info["item"][str(item_id)]
-    except KeyError:
-        item_name = "Item " + str(item_id) + " (Deprecated Item ID)"
-    return item_name
-
-
-def clean_timestamp(config_info, timestamp):
-    try:
-        age_days = (float(time.time()) - (float(timestamp) / 1000)) / 86400  # convert seconds to days
-    except ValueError:
-        age_days = "Unknown"
-    return age_days
+# def yvar_vs_count_in_lists(list_of_lists, y_var_list):
+#     # total_count = sum(sublist.count("TEAMMATENAME") for sublist in listoflists)
+#     return
 
 
 class Var:
